Migrating Arjun's code to XX_unary_arjun.go
